requirements.txt
# requirements.txt
# Python dependencies for Okta OIG API Management

# Core dependencies
requests>=2.31.0,<3.0.0

# OAG API Management (JWT authentication)
pyjwt>=2.8.0         # JWT token generation
cryptography>=41.0.0  # RSA key handling for OAG API

# Optional: For enhanced features
python-dotenv>=1.0.0  # Environment variable management
pyyaml>=6.0          # YAML configuration support
tabulate>=0.9.0      # Pretty-print tables in CLI
colorama>=0.4.6      # Colored terminal output

# Development dependencies (optional)
pytest>=7.4.0        # Testing framework
pytest-cov>=4.1.0    # Coverage reporting
black>=23.0.0        # Code formatting
flake8>=6.0.0        # Linting
mypy>=1.5.0          # Type checking
